1st text adventure I wrote from scratch in Ruby
There's not much of a story here, and I know there are some absurdly long lines in it that required a soft wrap. I have to learn how to more effectively do multi-line strings that are being called by a method. But I'm learning. Not bad for my first week of Ruby, I think!